+++ src/abstract/index.ts
@@ -24,11 +24,13 @@ export * from "./Key";
 export * from "./MainAxisAlignment";
 export * from "./MainAxisSize";
 export * from "./MouseCursor";
+export * from "./OpacityDecoration";
 export * from "./TextAlign";
 export * from "./TextDecoration";
 export * from "./TextDecorationStyle";
 export * from "./TextOverflow";
 export * from "./TextStyle";
 export * from "./VerticalDirection";
+
 // export * from './Width'
 // export * from './Height'