+++ src/abstract/index.ts
@@ -25,12 +25,10 @@ export * from "./MainAxisAlignment";
 export * from "./MainAxisSize";
 export * from "./MouseCursor";
 export * from "./OpacityDecoration";
+export * from "./SizedBox";
 export * from "./TextAlign";
 export * from "./TextDecoration";
 export * from "./TextDecorationStyle";
 export * from "./TextOverflow";
 export * from "./TextStyle";
 export * from "./VerticalDirection";
-
-// export * from './Width'
-// export * from './Height'