+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.3.0
 
 Add: ElevatedButton, TextButton, OpacityDecoration, Opacity Widget
+Add: SizedBox
+Add: ListView.builder with awesome (vue3-virtual-scroller!)[https://www.npmjs.com/package/vue3-virtual-scroller] which can be used in Flutter style now. ListView must be used inside SizedBox
 
 ## 0.2.0